Connect the AT91 UDC to USB CDC-ethernet support
This patch implements the low-level part of the USB CDC layer for AT91
Other boards can use this patch as an example to connect their own.

It does not enable the use of CDC ethernet for any of the AT91 boards.
Separate patches are required for that. Some boards also use network-
drivers that do not support CONFIG_NET_MULTI, which is mandatory for this
patch to work.

Signed-off-by: Thomas Smits <ts.smits@gmail.com>
Signed-off-by: Remy Bohmer <linux@bohmer.net>
Signed-off-by: Kyungmin Park <kyungmin.park@samsung.com>