samsung:misc: set_board_info: set a proper $partitions at boot
This change adds setting the environment variable: "partitions"
which is required to restore the partition table on the Trats2.

At this moment we are using MSDOS partition table on the Odroid
board, so for this board the above variable contains only a message.

Now the u-boot LCD menu in Trats2 is fully functional and the option:
(PWR+VOLUP) -> GPT restore - works well

Change-Id: I571466b7392a6775e0afe408c66be1dcc17d9ab8
Signed-off-by: Przemyslaw Marczak <p.marczak@samsung.com>