samsung: misc: add function for obtain the appropriate value according to the gpt version
This patch add get_gpt_environment() function in order to obtain
the appropriate value according to the gpt version.

This change introduces new function:
- get_gpt_environment() - Get some values according the the gpt version.

This function check "gpt_version" environment in order to check the current
gpt version. default version is ver 8 - for keep the support for old devices.
The default version can be changed after  added to allow partition change
through the sub-menu.

Odroid gpt partition information is to be added later.

Change-Id: I80c5389ad4947745b641487d9d785c1b86d8e94a
Signed-off-by: Inha Song <ideal.song@samsung.com>