fat: read: fix read issue
The present fat implementation ignores FAT16 long name directory entries
stored as the each block last entry. This causes that some files can't
be accessed by read in case if it's dir entries are stored in two blocks.
Now this is fixed by store the last used block in the buffer,
before read the next one.

Note that this is a temporary fix, since this breaks FAT32 read,
and will be fixed ASAP.

Change-Id: Ie2c2ccd07091d5fd09011c52d6f7c7dbdabd1dab
Signed-off-by: Przemyslaw Marczak <p.marczak@samsung.com>