bootstage: enable bootstage timing report to fdt
This commit enables boot stage time reporting for U-Boot
and appends it to device-tree, which can be seen in Linux.
The report is created as 'bootstage' fdt node with child
for each stage, like:
- main_loop (U-Boot prompt)
- start_kernel

fdt:
bootstage {
	44 {
		name = "start_kernel";
		mark = <>;
	};
	...
};

This can be checked in linux with 'cat /proc/device-tree/bootstage/id/name'.

Note: The id's are not constant for each boot stage name. For safety
use in linux, it's better to find the proper one by it's name.

This feature is enabled for configs:
- odroid-xu3
- odroid-u3
- tizen

Signed-off-by: Przemyslaw Marczak <p.marczak@samsung.com>